# dumps/serialize_to_writer: insert BufWriter to speed up serialization