@lschuermann 2024-10-21
@keith-hall @Enselic Thanks for the swift reviews! Funnily enough, it seems like this change has sped up syntect's own CI pipeline by 4-5x, hah!